move the profile image picker placeholder widgets into a different file
