+++ lib/widgets/custom/profile_image/profile_image_picker.dart
@@ -70,168 +70,3 @@ class ProfileImagePicker extends StatelessWidget {
     );
   }
 }
-
-/// This widget represents an action button for the button bar in a [ProfileImagePicker].
-///
-/// This button will use the minimum recommended size for interactive elements.
-class ProfileImagePickerActionButton extends StatelessWidget {
-  const ProfileImagePickerActionButton({
-    required this.androidIcon,
-    required this.iosIcon,
-    required this.onPressed,
-    this.isDestructive = false,
-    super.key,
-  });
-
-  /// The icon for the Android button.
-  final IconData androidIcon;
-
-  /// The icon for the iOS button.
-  final IconData iosIcon;
-
-  /// Whether this button represents a destructive action.
-  final bool isDestructive;
-
-  /// The onPressed handler for the button.
-  final void Function() onPressed;
-
-  @override
-  Widget build(BuildContext context) {
-    final theme = Theme.of(context);
-
-    Color backgroundColor;
-    IconData icon;
-    double size;
-
-    switch (theme.platform) {
-      case TargetPlatform.android:
-      case TargetPlatform.fuchsia:
-      case TargetPlatform.linux:
-      case TargetPlatform.windows:
-        backgroundColor =
-            isDestructive ? theme.colorScheme.error : theme.colorScheme.primary;
-        icon = androidIcon;
-        size = kMinInteractiveDimension;
-        break;
-      case TargetPlatform.iOS:
-      case TargetPlatform.macOS:
-        backgroundColor = isDestructive
-            ? CupertinoColors.destructiveRed
-            : CupertinoTheme.of(context).primaryColor;
-        icon = iosIcon;
-        size = kMinInteractiveDimensionCupertino;
-        break;
-    }
-
-    final decorationSize = size * 0.75;
-
-    return GestureDetector(
-      behavior: HitTestBehavior.opaque,
-      onTap: onPressed,
-      child: SizedBox.square(
-        dimension: size,
-        child: Center(
-          child: Container(
-            width: decorationSize,
-            height: decorationSize,
-            decoration: BoxDecoration(
-              color: backgroundColor,
-              borderRadius: BorderRadius.all(
-                Radius.circular(decorationSize / 2),
-              ),
-            ),
-            child: Center(
-              child: Icon(
-                icon,
-                color: Colors.white,
-                size: decorationSize * 0.6,
-              ),
-            ),
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-/// This widget represents a placeholder for the [ProfileImagePicker]
-/// that displays an iOS-style grey circular shape
-/// with a clipped person icon in the middle.
-class _CupertinoProfileImagePickerPlaceholder extends StatelessWidget {
-  const _CupertinoProfileImagePickerPlaceholder({required this.size});
-
-  /// The size for the placeholder.
-  final double size;
-
-  @override
-  Widget build(BuildContext context) {
-    final borderWidth = size / 16;
-
-    return SizedBox.square(
-      dimension: size,
-      child: Stack(
-        children: [
-          Container(
-            width: size,
-            height: size,
-            decoration: BoxDecoration(
-              color: CupertinoColors.systemGrey2,
-              borderRadius: BorderRadius.all(
-                Radius.circular(size / 2),
-              ),
-            ),
-          ),
-          Align(
-            alignment: Alignment.bottomCenter,
-            child: Transform.translate(
-              offset: Offset(0, borderWidth),
-              child: Icon(
-                CupertinoIcons.person_solid,
-                color: Colors.white,
-                size: size,
-              ),
-            ),
-          ),
-          Container(
-            width: size,
-            height: size,
-            decoration: BoxDecoration(
-              border: Border.all(
-                color: CupertinoColors.systemGrey2,
-                width: borderWidth,
-              ),
-              borderRadius: BorderRadius.all(
-                Radius.circular(size / 2),
-              ),
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-}
-
-/// This widget represents a placeholder for the [ProfileImagePicker]
-/// that displays a Material-style circular shape
-/// with a person icon in the middle.
-class _MaterialProfileImagePickerPlaceholder extends StatelessWidget {
-  const _MaterialProfileImagePickerPlaceholder({required this.size});
-
-  /// The size of the placeholder.
-  final double size;
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      height: size,
-      width: size,
-      decoration: BoxDecoration(
-        borderRadius: BorderRadius.all(Radius.circular(size / 2)),
-        color: Colors.grey.shade500,
-      ),
-      child: Center(
-        child: Icon(Icons.person, color: Colors.white, size: .7 * size),
-      ),
-    );
-  }
-}
